Update script for SC23 titleslide generation
Adapt to new presentations and to new SC23 theme. The TeX template now uses TikZ to match larger titles. The Makefile for generating the individual Slide PDFs now has an inserted pdfjam step to match the papersize of the TeX-generated SC23 title slide and the other slides (which might have gigantous papersizes if they are created by PowerPoint).